基础视频播放

+++ BaseProject/BaseProject/Classes/Vendors/Player/ZBData.hpp
@@ -21,7 +21,7 @@ enum ZBDataType
 struct ZBData
 {
     int type = 0;
-    int pts = 0;
+    long long pts = 0;
     unsigned char *data = 0;
     unsigned char *datas[8] = {0};//Decoded data
     int size = 0;
